+++ noodles-tabix/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.21.0 - 2023-05-18
 
 ### Added